controle de travas - multi abas

+++ nds-client/src/main/java/br/com/abril/nds/client/component/BloqueioConferenciaEncalheComponent.java
@@ -1,5 +1,12 @@
 package br.com.abril.nds.client.component;
 
+import java.util.HashMap;
+import java.util.Map;
+
+
+
+
+
 import javax.servlet.http.HttpSession;
 
 import org.slf4j.Logger;
@@ -16,6 +23,7 @@ import br.com.abril.nds.model.seguranca.ConferenciaEncalheCotaUsuario;
 import br.com.abril.nds.model.seguranca.Usuario;
 import br.com.abril.nds.repository.ConferenciaEncalheCotaUsuarioRepository;
 import br.com.abril.nds.service.UsuarioService;
+import br.com.abril.nds.vo.ValidacaoVO;
 
 @Component
 public class BloqueioConferenciaEncalheComponent {
@@ -27,7 +35,8 @@ public class BloqueioConferenciaEncalheComponent {
 	
 	@Autowired
 	private ConferenciaEncalheCotaUsuarioRepository conferenciaEncalheCotaUsuarioRepository; 
-
+	
+	
 	/**
 	 * Se o usuário (session id) ja estiver conferindo alguma cota,
 	 * não sera permitido que este inicie uma nova conferência.
@@ -40,10 +49,19 @@ public class BloqueioConferenciaEncalheComponent {
 		String user = usuarioService.getUsuarioLogado().getLogin();
 		
 		ConferenciaEncalheCotaUsuario cuc = conferenciaEncalheCotaUsuarioRepository.obterPorNumeroCota(numeroCota);
-			
+		if ( cuc != null ) {
+		String windowname = (String) session.getAttribute("WINDOWNAME");
+		String windowname_conferencia = (String) session.getAttribute("WINDOWNAME_CONFERENCIA");
+		if ( windowname != null && windowname_conferencia != null && !windowname.equals(windowname_conferencia)) {
+			throw new ValidacaoException(TipoMensagem.WARNING, String.format("Não é possível executar Conferência de Encalhe simultânea para o usuário em mais que uma aba: %s!", cuc.getLogin()));
+		}	
+		}
 		if(cuc == null || (cuc.getLogin().equals(user) && cuc.getSessionId().equals(session.getId()) && cuc.getNumeroCota().equals(numeroCota))) {
 			return;
 		}
+		
+		
+		
 		
 		if(cuc.getLogin().equals(user) && !cuc.getSessionId().equals(session.getId())) {
 
@@ -85,7 +103,29 @@ public class BloqueioConferenciaEncalheComponent {
 	@Transactional
 	public void removerTravaConferenciaCotaUsuario(HttpSession session) {
 		
-		conferenciaEncalheCotaUsuarioRepository.removerPorSessionId(session.getId());
+	
+       // remover bloqueio de acordo com o sessionid de quando o bloqueio foi criado
+		String windowname = (String) session.getAttribute("WINDOWNAME");
+		String windowname_conferencia = (String) session.getAttribute("WINDOWNAME_CONFERENCIA");
+		if ( windowname != null && windowname_conferencia != null && !windowname.equals(windowname_conferencia)) {
+			//  nao liberar , pois nao he mesma janela
+			return;
+		}	
+		 windowname = (String) session.getAttribute("WINDOWNAME_CONFERENCIA");
+    		  if (windowname != null ) {
+    			  Map map =  (Map) session.getAttribute("WINDOWNAME_MAP");
+    			  if ( map != null ) {
+    				  String sessionid =(String) map.get(windowname);
+    				  if ( sessionid != null ) {
+    					 if ( conferenciaEncalheCotaUsuarioRepository.removerPorSessionId(sessionid) > 0 )
+    						  return ; // removeu 
+    				  }
+    			  }
+    			  }
+    		// tentar remover pelo sessoinid atual, que pode nao ser o usuado na trava, se o usuario abriu mais de uma aba e fez login na outra aba
+    			conferenciaEncalheCotaUsuarioRepository.removerPorSessionId(session.getId());
+    		  
+    		  
 	}
 	
 	public String getIdentificacaoUnicaUsuarioLogado(HttpSession session) {
@@ -112,17 +152,51 @@ public class BloqueioConferenciaEncalheComponent {
 		
 		if(!existeTravaConferenciaCotaUsuario(numeroCota, session)) {
 
+			
 			conferenciaEncalheCotaUsuarioRepository.removerPorLogin(login);
 			ConferenciaEncalheCotaUsuario cuc = new ConferenciaEncalheCotaUsuario(login, nomeUsuario, session.getId(), numeroCota);
 			try {
-				
+				LOGGER.error("INSERINDO TRAVA NA TABELA CONFERENCIA_ENCALHE_COTA_USUARIO");
 				conferenciaEncalheCotaUsuarioRepository.adicionar(cuc);
+				// salvar qual janela fez o bloqueio para remove-lo para recuperar o sessionid usado no bloqueio
+				// para evitar o problema de multiplas abas, onde o sessionid pode mudar
+				String windowname = (String) session.getAttribute("WINDOWNAME"); 
+				if ( windowname != null ) {
+				session.setAttribute("WINDOWNAME_CONFERENCIA",windowname);
+				// salvar windowname ( setado no loginfilter vindo do cookie WINDOWNAME 
+            	// para evitar usuario logado em abas diferences acessar estudo/encalhe/etc mais de uma vez
+       			  // salvar sessao associado com o map
+    			  Map map = (Map) session.getAttribute("WINDOWNAME_MAP");
+    			  if ( map == null ) {
+    				   map = new HashMap();
+    				   map.put(windowname, session.getId());
+    			  } else {
+    				  if ( !map.containsKey(windowname))
+    						  map.put(windowname,session.getId());
+    			  }
+    			  session.setAttribute("WINDOWNAME_MAP",map);
+    			  }
+				
+				// conferir se ja nao tem um estudo aberto em outra aba nesta sessao
+		    	String windowname_ce=(String) session.getAttribute("WINDOWNAME_CONFERENCIA");
+		    	String windowname1=(String) session.getAttribute("WINDOWNAME");
+		    	
+		      	 
+		    	if ( windowname_ce != null && windowname1 != null && !windowname_ce.equals(windowname1))
+		    	{
+		    		 throw new ValidacaoException(new ValidacaoVO(TipoMensagem.WARNING, "Ja existe uma Conferencia sendo executada em outra aba/janela"));
+		    	}
+		    	
+				
+				
 			} catch(Throwable e) {
 				
 				LOGGER.error("A Cota %s está sendo conferida.", e);
 				throw new ValidacaoException(TipoMensagem.WARNING, String.format("A Cota %s está sendo conferida.", numeroCota));
 			}
 		}
+		
+		
 	}
 
 	@Transactional
